+++ src/model/Item.java
@@ -7,6 +7,6 @@ public interface Item {
 	
 	public String getName();
 	public double getPrice();
-	public double getCalories();
+	public int getCalories();
 
 }
